ChartBot-Pro-Refined/requirements.txt
PyQt5
pytesseract
Pillow
